Add warning for clusters represented by a single taxon
We try to add a warning to the user in case a cluster is represented by a single taxon.
Since this is usually ok, we set a default option to not show the warning, so, for now,
the user must explicitly state check_singles = TRUE. This default option also makes
the testing easier.