Correct links and update description
There is an issue with the doi link that needs to be corrected, so the package is failing R CMD check --as-cran while the doi is not working.
In the DESCRIPTION file there is a problem with a cran link that needs to be in canonical form and there is an unnecessary line.