Add additional checks to user input data.
We try to solve a problem, which is the fact that some users might have species abundance tables that contain
samples with very few species, including a single species.

In this scenarios, ulrb will not work as expected.

At this stage, we are adding additional checks to define_rb() function, so that it can identify such samples
and provide a warning -- we are discarding sample with 1 or 2 species.

This was not tested yet.

We also need to incorporate additional checks for the automated version.