Ensure that the user uses the correct number of clusters
Add steps to verify the maximum number of k (species number -1) possible to calculate in each sample.
Samples with 3 or less species are discarded as of this version (0.1.8), because they are not appropriate for the clustering algorithm.

Error messages were updated to indicate the user what is the necessary range of k values (in case the automatic parameter is set to TRUE).

Note: I'm aware that the new steps to verify if the range of k values selected by the user are appropriate is very inneficient, but right now I don't have time to
add these steps in a more efficient way. I will try to improve this in the future.